combine {RRF}	R Documentation

Combine Ensembles of Trees

Description

Combine two more more ensembles of trees into one.

Usage

combine(...)

Arguments

...
two or more objects of class RRF, to be combined into one.

Value

An object of class RRF.

Note

The confusion, err.rate, mse and rsq components (as well as the corresponding components in the test compnent, if exist) of the combined object will be NULL.

Author(s)

Andy Liaw andy\_liaw@merck.com

See Also

RRF, grow

Examples

data(iris)
rf1 <- RRF(Species ~ ., iris, ntree=50, norm.votes=FALSE)
rf2 <- RRF(Species ~ ., iris, ntree=50, norm.votes=FALSE)
rf3 <- RRF(Species ~ ., iris, ntree=50, norm.votes=FALSE)
rf.all <- combine(rf1, rf2, rf3)
print(rf.all)

[Package RRF version 1.9 Index]