grow {RRF}	R Documentation

Add trees to an ensemble

Description

Add additional trees to an existing ensemble of trees.

Usage

## S3 method for class 'RRF'
grow(x, how.many, ...)

Arguments

x
an object of class RRF, which contains a forest component.

how.many
number of trees to add to the RRF object.

...
currently ignored.

Value

An object of class RRF, containing how.many additional trees.

Note

The confusion, err.rate, mse and rsq components (as well as the corresponding components in the test compnent, if exist) of the combined object will be NULL.

Author(s)

Andy Liaw andy\_liaw@merck.com

See Also

combine, RRF

Examples

data(iris)
iris.rf <- RRF(Species ~ ., iris, ntree=50, norm.votes=FALSE)
iris.rf <- grow(iris.rf, 50)
print(iris.rf)

[Package RRF version 1.8 Index]