margin {RRF}	R Documentation

Margins of RRF Classifier

Description

Compute or plot the margin of predictions from a RRF classifier.

Usage

## S3 method for class 'RRF'
margin(x, ...)
## Default S3 method:
margin(x, observed, ...)
## S3 method for class 'margin'
plot(x, sort=TRUE, ...)

Arguments

x
an object of class RRF, whose type is not regression, or a matrix of predicted probabilities, one column per class and one row per observation. For the plot method, x should be an object returned by margin.

observed
the true response corresponding to the data in x.

sort
Should the data be sorted by their class labels?

...
other graphical parameters to be passed to plot.default.

Value

For margin, the margin of observations from the RRF classifier (or whatever classifier that produced the predicted probability matrix given to margin). The margin of a data point is defined as the proportion of votes for the correct class minus maximum proportion of votes for the other classes. Thus under majority votes, positive margin means correct classification, and vice versa.

Author(s)

Robert Gentlemen, with slight modifications by Andy Liaw

See Also

RRF

Examples

set.seed(1)
data(iris)
iris.rf <- RRF(Species ~ ., iris, keep.forest=FALSE)
plot(margin(iris.rf))

[Package RRF version 1.9 Index]