na.roughfix {RRF}	R Documentation

Rough Imputation of Missing Values

Description

Impute Missing Values by median/mode.

Usage

na.roughfix(object, ...)

Arguments

object
a data frame or numeric matrix.

...
further arguments special methods could require.

Value

A completed data matrix or data frame. For numeric variables, NAs are replaced with column medians. For factor variables, NAs are replaced with the most frequent levels (breaking ties at random). If object contains no NAs, it is returned unaltered.

Note

This is used as a starting point for imputing missing values by random forest.

Author(s)

Andy Liaw

See Also

rrfImpute, RRF.

Examples

data(iris)
iris.na <- iris
set.seed(111)
## artificially drop some data values.
for (i in 1:4) iris.na[sample(150, sample(20)), i] <- NA
iris.roughfix <- na.roughfix(iris.na)
iris.narf <- RRF(Species ~ ., iris.na, na.action=na.roughfix)
print(iris.narf)

[Package RRF version 1.8 Index]