outlier {RRF}	R Documentation

Compute outlying measures

Description

Compute outlying measures based on a proximity matrix.

Usage

## Default S3 method:
outlier(x, cls=NULL, ...)
## S3 method for class 'RRF'
outlier(x, ...)

Arguments

x
a proximity matrix (a square matrix with 1 on the diagonal and values between 0 and 1 in the off-diagonal positions); or an object of class RRF, whose type is not regression.

cls
the classes the rows in the proximity matrix belong to. If not given, all data are assumed to come from the same class.

...
arguments for other methods.

Value

A numeric vector containing the outlying measures. The outlying measure of a case is computed as n / sum(squared proximity), normalized by subtracting the median and divided by the MAD, within each class.

See Also

RRF

Examples

set.seed(1)
iris.rf <- RRF(iris[,-5], iris[,5], proximity=TRUE)
plot(outlier(iris.rf), type="h",
     col=c("red", "green", "blue")[as.numeric(iris$Species)])

[Package RRF version 1.9 Index]