plot.RRF {RRF}	R Documentation

Plot method for RRF objects

Description

Plot the error rates or MSE of a RRF object

Usage

## S3 method for class 'RRF'
plot(x, type="l", main=deparse(substitute(x)), ...)

Arguments

x
an object of class RRF.

type
type of plot.

main
main title of the plot.

...
other graphical parameters.

Value

Invisibly, the error rates or MSE of the RRF object. If the object has a non-null test component, then the returned object is a matrix where the first column is the out-of-bag estimate of error, and the second column is for the test set.

Note

This function does not work for RRF objects that have type=unsupervised.

If the x has a non-null test component, then the test set errors are also plotted.

Author(s)

Andy Liaw

See Also

RRF

Examples

data(mtcars)
plot(RRF(mpg ~ ., mtcars, keep.forest=FALSE, ntree=100), log="y")

[Package RRF version 1.9 Index]