varImpPlot {RRF}	R Documentation

Variable Importance Plot

Description

Dotchart of variable importance as measured by a Random Forest

Usage

varImpPlot(x, sort=TRUE, n.var=min(30, nrow(x$importance)),
           type=NULL, class=NULL, scale=TRUE, 
           main=deparse(substitute(x)), ...) 

Arguments

x
An object of class RRF.

sort
Should the variables be sorted in decreasing order of importance?

n.var
How many variables to show? (Ignored if sort=FALSE.)

type, class, scale
arguments to be passed on to importance

main
plot title.

...
Other graphical parameters to be passed on to dotchart.

Value

Invisibly, the importance of the variables that were plotted.

Author(s)

Andy Liaw andy\_liaw@merck.com

See Also

RRF, importance

Examples

set.seed(4543)
data(mtcars)
mtcars.rf <- RRF(mpg ~ ., data=mtcars, ntree=1000, keep.forest=FALSE,
                          importance=TRUE)
varImpPlot(mtcars.rf)

[Package RRF version 1.8 Index]