varUsed {RRF}	R Documentation

Variables used in a random forest

Description

Find out which predictor variables are actually used in the random forest.

Usage

varUsed(x, by.tree=FALSE, count=TRUE)

Arguments

x
An object of class RRF.

by.tree
Should the list of variables used be broken down by trees in the forest?

count
Should the frequencies that variables appear in trees be returned?

Value

If count=TRUE and by.tree=FALSE, a integer vector containing frequencies that variables are used in the forest. If by.tree=TRUE, a matrix is returned, breaking down the counts by tree (each column corresponding to one tree and each row to a variable).

If count=FALSE and by.tree=TRUE, a list of integer indices is returned giving the variables used in the trees, else if by.tree=FALSE, a vector of integer indices giving the variables used in the entire forest.

Author(s)

Andy Liaw

See Also

RRF

Examples

data(iris)
set.seed(17)
varUsed(RRF(Species~., iris, ntree=100))

[Package RRF version 1.8 Index]